utils/requirements.txt
pandas
flask
tweepy
tqdm
json
ibm-watson
pyyaml
